+++ README.md
@@ -1,4 +1,4 @@
-# otus spring homework 9
-homework 9 for Otus springFramework course
+# otus spring homework 10
+homework 10 for Otus springFramework course
 
-CRUD app with Web UI (Thymeleaf, classic controllers)
+AJAX and REST
