+++ pom.xml
@@ -19,7 +19,7 @@
     </parent>
 
     <properties>
-        <homework.ver>9</homework.ver>
+        <homework.ver>10</homework.ver>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <java.version>1.8</java.version>
