+++ src/main/java/ru/ezikvice/springotus/homework10/controller/BooksController.java
@@ -1,4 +1,4 @@
-package ru.ezikvice.springotus.homework9.controller;
+package ru.ezikvice.springotus.homework10.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -6,13 +6,13 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import ru.ezikvice.springotus.homework9.domain.Author;
-import ru.ezikvice.springotus.homework9.domain.Book;
-import ru.ezikvice.springotus.homework9.domain.Comment;
-import ru.ezikvice.springotus.homework9.domain.Genre;
-import ru.ezikvice.springotus.homework9.service.AuthorService;
-import ru.ezikvice.springotus.homework9.service.BookService;
-import ru.ezikvice.springotus.homework9.service.GenreService;
+import ru.ezikvice.springotus.homework10.domain.Author;
+import ru.ezikvice.springotus.homework10.domain.Book;
+import ru.ezikvice.springotus.homework10.domain.Comment;
+import ru.ezikvice.springotus.homework10.domain.Genre;
+import ru.ezikvice.springotus.homework10.service.AuthorService;
+import ru.ezikvice.springotus.homework10.service.BookService;
+import ru.ezikvice.springotus.homework10.service.GenreService;
 
 import java.util.List;
