+++ src/main/java/ru/ezikvice/springotus/homework10/dao/AuthorRepository.java
@@ -1,7 +1,7 @@
-package ru.ezikvice.springotus.homework9.dao;
+package ru.ezikvice.springotus.homework10.dao;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import ru.ezikvice.springotus.homework9.domain.Author;
+import ru.ezikvice.springotus.homework10.domain.Author;
 
 import java.util.Set;
