+++ src/main/java/ru/ezikvice/springotus/homework10/dao/GenreRepository.java
@@ -1,7 +1,7 @@
-package ru.ezikvice.springotus.homework9.dao;
+package ru.ezikvice.springotus.homework10.dao;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import ru.ezikvice.springotus.homework9.domain.Genre;
+import ru.ezikvice.springotus.homework10.domain.Genre;
 
 public interface GenreRepository extends MongoRepository<Genre, Long> {
     Genre save(Genre genre);
