+++ src/main/java/ru/ezikvice/springotus/homework10/service/BookService.java
@@ -1,9 +1,9 @@
-package ru.ezikvice.springotus.homework9.service;
+package ru.ezikvice.springotus.homework10.service;
 
-import ru.ezikvice.springotus.homework9.domain.Author;
-import ru.ezikvice.springotus.homework9.domain.Book;
-import ru.ezikvice.springotus.homework9.domain.Comment;
-import ru.ezikvice.springotus.homework9.domain.Genre;
+import ru.ezikvice.springotus.homework10.domain.Author;
+import ru.ezikvice.springotus.homework10.domain.Book;
+import ru.ezikvice.springotus.homework10.domain.Comment;
+import ru.ezikvice.springotus.homework10.domain.Genre;
 
 import java.util.List;
 import java.util.Set;
