+++ src/main/java/ru/ezikvice/springotus/homework10/service/BookServiceImpl.java
@@ -1,14 +1,14 @@
-package ru.ezikvice.springotus.homework9.service;
+package ru.ezikvice.springotus.homework10.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import ru.ezikvice.springotus.homework9.dao.AuthorRepository;
-import ru.ezikvice.springotus.homework9.dao.BookRepository;
-import ru.ezikvice.springotus.homework9.dao.GenreRepository;
-import ru.ezikvice.springotus.homework9.domain.Author;
-import ru.ezikvice.springotus.homework9.domain.Book;
-import ru.ezikvice.springotus.homework9.domain.Comment;
-import ru.ezikvice.springotus.homework9.domain.Genre;
+import ru.ezikvice.springotus.homework10.dao.AuthorRepository;
+import ru.ezikvice.springotus.homework10.dao.BookRepository;
+import ru.ezikvice.springotus.homework10.dao.GenreRepository;
+import ru.ezikvice.springotus.homework10.domain.Author;
+import ru.ezikvice.springotus.homework10.domain.Book;
+import ru.ezikvice.springotus.homework10.domain.Comment;
+import ru.ezikvice.springotus.homework10.domain.Genre;
 
 import java.util.List;
 import java.util.Set;
