+++ src/main/java/ru/ezikvice/springotus/homework10/service/GenreServiceImpl.java
@@ -1,9 +1,9 @@
-package ru.ezikvice.springotus.homework9.service;
+package ru.ezikvice.springotus.homework10.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import ru.ezikvice.springotus.homework9.dao.GenreRepository;
-import ru.ezikvice.springotus.homework9.domain.Genre;
+import ru.ezikvice.springotus.homework10.dao.GenreRepository;
+import ru.ezikvice.springotus.homework10.domain.Genre;
 
 import java.util.List;
