+++ src/test/java/ru/ezikvice/springotus/homework10/AuthorDaoTests.java
@@ -1,4 +1,4 @@
-package ru.ezikvice.springotus.homework9;
+package ru.ezikvice.springotus.homework10;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -8,30 +8,27 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.test.context.junit4.SpringRunner;
-import ru.ezikvice.springotus.homework9.dao.BookRepository;
-import ru.ezikvice.springotus.homework9.domain.Book;
-import ru.ezikvice.springotus.homework9.domain.Comment;
+import ru.ezikvice.springotus.homework10.dao.AuthorRepository;
+import ru.ezikvice.springotus.homework10.domain.Author;
 
 import static org.junit.Assert.assertTrue;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
 @DataMongoTest
-@ComponentScan("ru.ezikvice.springotus.homework9.dao")
-public class CommentDaoTests {
+@ComponentScan("ru.ezikvice.springotus.homework10.dao")
+public class AuthorDaoTests {
 
     @Autowired
     private MongoTemplate mt;
 
     @Autowired
-    private BookRepository rep;
+    private AuthorRepository rep;
 
     @Test
-    public void savingCommentToBook() {
-        Book book = new Book("Very Popular Book", "Book with comments");
-        Comment comment = new Comment("Test Name Author");
-        book.addComment(comment);
-        rep.save(book);
-        assertTrue(book.getComments().contains(comment));
+    public void findingAuthorByName() {
+        Author testNameAuthor = new Author("Test Name Author");
+        mt.save(testNameAuthor);
+        assertTrue(rep.findByName("Test Name Author").contains(testNameAuthor));
     }
 }
