+++ src/test/java/ru/ezikvice/springotus/homework10/BookTests.java
@@ -1,16 +1,16 @@
-package ru.ezikvice.springotus.homework9;
+package ru.ezikvice.springotus.homework10;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
-import ru.ezikvice.springotus.homework9.domain.Author;
-import ru.ezikvice.springotus.homework9.domain.Book;
-import ru.ezikvice.springotus.homework9.domain.Genre;
-import ru.ezikvice.springotus.homework9.service.AuthorService;
-import ru.ezikvice.springotus.homework9.service.BookService;
-import ru.ezikvice.springotus.homework9.service.GenreService;
+import ru.ezikvice.springotus.homework10.domain.Author;
+import ru.ezikvice.springotus.homework10.domain.Book;
+import ru.ezikvice.springotus.homework10.domain.Genre;
+import ru.ezikvice.springotus.homework10.service.AuthorService;
+import ru.ezikvice.springotus.homework10.service.BookService;
+import ru.ezikvice.springotus.homework10.service.GenreService;
 
 import java.util.Set;
